Extract AbstractKieRepositoryScanner as an extension point for alternative, non-Maven repository based scanners.
